Configure pageExtensions and rename page files to .page.tsx to prevent Next.js from trying to treat other files in the directory as pages
